♻️ [refactor]: #452 Phase 9 basic: lexer_read_real_basic.rs を分離
read_real の基本テスト（Phase 9-D / 9-B / 9-C / 9-D続き / 9-E / 9-F / 9-A）を
`src/lexer/tests/lexer_read_real_basic.rs` に分離。実コード上のサブ Phase は
意味順ではないため、行範囲（旧 lexer.rs:1577-2042）= Phase 9 ヘッダから
Phase 9-G サブヘッダ手前までを境界として確定。

- 293 件グリーン維持
- 短縮テスト関数名集合不変 / 重複 0
- lexer_read_real_basic.rs: 467 行 (< 800)

Refs #452
